Fix password parameter in UserCreateService and null check in CriteriaPrimitiveToCriteriaMapper
The password parameter in the createUser method of UserCreateService has been corrected to use the correct method parameter instead of userEmail. Additionally, a null check has been added in CriteriaPrimitiveToCriteriaMapper before mapping filters to prevent NullPointerException if the filters list is null. Also, a minor cleanup was made in the CriteriaRepositoryImpl class and a typo in the api-tests.http file has been corrected.